+++ src/main/groovy/netflix/nebula/dependency/recommender/provider/MavenBomRecommendationProvider.java
@@ -1,7 +1,24 @@
+/*
+ * Copyright 2014-2017 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package netflix.nebula.dependency.recommender.provider;
 
+import com.netflix.nebula.dependencybase.DependencyManagement;
 import org.apache.maven.model.Dependency;
 import org.apache.maven.model.Model;
+import org.apache.maven.model.Parent;
 import org.apache.maven.model.Repository;
 import org.apache.maven.model.building.*;
 import org.apache.maven.model.interpolation.StringSearchModelInterpolator;
@@ -23,6 +40,7 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
+import java.net.URI;
 import java.net.URL;
 import java.net.URLConnection;
 import java.util.*;
@@ -30,12 +48,19 @@ import org.gradle.api.artifacts.repositories.PasswordCredentials;
 
 public class MavenBomRecommendationProvider extends ClasspathBasedRecommendationProvider {
     private Map<String, String> recommendations;
+    private DependencyManagement insight;
 
     public MavenBomRecommendationProvider(Project project, String configName) {
         super(project, configName);
+        this.insight = new DependencyManagement();
     }
 
-    private static class SimpleModelSource implements ModelSource {
+    public MavenBomRecommendationProvider(Project project, String configName, DependencyManagement insight) {
+        super(project, configName);
+        this.insight = insight;
+    }
+
+    private static class SimpleModelSource implements ModelSource2 {
         InputStream in;
 
         public SimpleModelSource(InputStream in) {
@@ -49,6 +74,16 @@ public class MavenBomRecommendationProvider extends ClasspathBasedRecommendation
 
         @Override
         public String getLocation() {
+            return null;
+        }
+
+        @Override
+        public ModelSource2 getRelatedSource(String relPath) {
+            return null;
+        }
+
+        @Override
+        public URI getLocationURI() {
             return null;
         }
     }
@@ -68,7 +103,7 @@ public class MavenBomRecommendationProvider extends ClasspathBasedRecommendation
 
                 request.setModelResolver(new ModelResolver() {
                     @Override
-                    public ModelSource resolveModel(String groupId, String artifactId, String version) throws UnresolvableModelException {
+                    public ModelSource2 resolveModel(String groupId, String artifactId, String version) throws UnresolvableModelException {
                         String relativeUrl = buildRelativeUrl(groupId, artifactId, version);
                         List<String> repositoryErrors = new LinkedList<>();
                         try {
@@ -93,7 +128,22 @@ public class MavenBomRecommendationProvider extends ClasspathBasedRecommendation
                     }
 
                     @Override
+                    public ModelSource2 resolveModel(Dependency dependency) throws UnresolvableModelException {
+                        return resolveModel(dependency.getGroupId(), dependency.getArtifactId(), dependency.getVersion());
+                    }
+
+                    @Override
+                    public ModelSource2 resolveModel(Parent parent) throws UnresolvableModelException {
+                        return resolveModel(parent.getGroupId(), parent.getArtifactId(), parent.getVersion());
+                    }
+
+                    @Override
                     public void addRepository(Repository repository) throws InvalidRepositoryException {
+                        // do nothing
+                    }
+
+                    @Override
+                    public void addRepository(Repository repository, boolean bool) throws InvalidRepositoryException {
                         // do nothing
                     }
 
@@ -130,11 +180,13 @@ public class MavenBomRecommendationProvider extends ClasspathBasedRecommendation
                     }
                 });
                 request.setModelSource(new SimpleModelSource(new FileInputStream(recommendation)));
+                request.setSystemProperties(System.getProperties());
 
                 DefaultModelBuilder modelBuilder = new DefaultModelBuilderFactory().newInstance();
                 modelBuilder.setModelInterpolator(new ProjectPropertiesModelInterpolator(project));
 
                 ModelBuildingResult result = modelBuilder.build(request);
+                insight.addPluginMessage("nebula.dependency-recommender uses mavenBom: " + result.getEffectiveModel().getId());
                 for (Dependency d : result.getEffectiveModel().getDependencyManagement().getDependencies()) {
                     recommendations.put(d.getGroupId() + ":" + d.getArtifactId(), d.getVersion());
                 }