+++ src/test/groovy/netflix/nebula/dependency/recommender/DependencyRecommendationsPluginSpec.groovy
@@ -1,16 +1,33 @@
+/*
+ * Copyright 2016-2017 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package netflix.nebula.dependency.recommender
 
 import nebula.test.IntegrationSpec
 import nebula.test.dependencies.DependencyGraphBuilder
-import nebula.test.dependencies.DependencyGraphNode
 import nebula.test.dependencies.GradleDependencyGenerator
 import nebula.test.dependencies.ModuleBuilder
 import nebula.test.dependencies.maven.ArtifactType
 import nebula.test.dependencies.maven.Pom
 import nebula.test.dependencies.repositories.MavenRepo
+import org.gradle.api.artifacts.ResolvedDependency
+import org.gradle.api.artifacts.result.ResolvedDependencyResult
 import org.gradle.testfixtures.ProjectBuilder
 import org.xmlunit.builder.DiffBuilder
 import org.xmlunit.builder.Input
+import spock.lang.Issue
 
 class DependencyRecommendationsPluginSpec extends IntegrationSpec  {
     def 'applies recommendations to dependencies with no version'() {
@@ -76,6 +93,42 @@ class DependencyRecommendationsPluginSpec extends IntegrationSpec  {
 
         then:
         result.standardOutput.contains 'test.nebula:foo: -> 1.0.0'
+    }
+
+    def 'dependencyInsightEnhanced from recommendation via configuration'() {
+        def repo = new MavenRepo()
+        repo.root = new File(projectDir, 'build/bomrepo')
+        def pom = new Pom('test.nebula.bom', 'testbom', '1.0.0', ArtifactType.POM)
+        pom.addManagementDependency('test.nebula', 'foo', '1.0.0')
+        repo.poms.add(pom)
+        repo.generate()
+        def graph = new DependencyGraphBuilder()
+                .addModule('test.nebula:foo:1.0.0')
+                .build()
+        def generator = new GradleDependencyGenerator(graph)
+        generator.generateTestMavenRepo()
+
+        buildFile << """\
+            apply plugin: 'nebula.dependency-recommender'
+            apply plugin: 'java'
+            
+            repositories {
+                maven { url '${repo.root.absolutePath}' }
+                ${generator.mavenRepositoryBlock}
+            }
+            
+            dependencies {
+                nebulaRecommenderBom 'test.nebula.bom:testbom:1.0.0@pom'
+                compile 'test.nebula:foo'
+            }
+            """.stripIndent()
+
+        when:
+        def result = runTasksSuccessfully('dependencyInsightEnhanced', '--configuration', 'compile', '--dependency' , 'foo')
+
+        then:
+        result.standardOutput.contains 'test.nebula:foo:1.0.0 (recommend 1.0.0 via conflict resolution recommendation)'
+        result.standardOutput.contains 'nebula.dependency-recommender uses mavenBom: test.nebula.bom:testbom:pom:1.0.0'
     }
 
     def 'conflict resolved respects higher transitive'() {
@@ -249,63 +302,102 @@ class DependencyRecommendationsPluginSpec extends IntegrationSpec  {
         def diff = DiffBuilder
                 .compare(Input.fromString(pomText))
                 .withTest(Input.fromString('''\
-                    <project xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-                      <modelVersion>4.0.0</modelVersion>
-                      <groupId>netflix</groupId>
-                      <artifactId>module-parent</artifactId>
-                      <version>1</version>
-                      <packaging>pom</packaging>
-                      <dependencyManagement>
-                        <dependencies>
-                          <dependency>
-                            <groupId>commons-digester</groupId>
-                            <artifactId>commons-digester</artifactId>
-                            <version>1.8</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-logging</groupId>
-                            <artifactId>commons-logging</artifactId>
-                            <version>1.1.1</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-lang</groupId>
-                            <artifactId>commons-lang</artifactId>
-                            <version>2.4</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-configuration</groupId>
-                            <artifactId>commons-configuration</artifactId>
-                            <version>1.6</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-beanutils</groupId>
-                            <artifactId>commons-beanutils</artifactId>
-                            <version>1.7.0</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-collections</groupId>
-                            <artifactId>commons-collections</artifactId>
-                            <version>3.2.1</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>commons-beanutils</groupId>
-                            <artifactId>commons-beanutils-core</artifactId>
-                            <version>1.8.0</version>
-                          </dependency>
-                          <dependency>
-                            <groupId>manual</groupId>
-                            <artifactId>dep</artifactId>
-                            <version>1</version>
-                          </dependency>
-                        </dependencies>
-                      </dependencyManagement>
-                      <description>A demonstration of maven POM customization</description>
-                    </project>
+                        <project xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+                          <modelVersion>4.0.0</modelVersion>
+                          <groupId>netflix</groupId>
+                          <artifactId>module-parent</artifactId>
+                          <version>1</version>
+                          <packaging>pom</packaging>
+                          <dependencyManagement>
+                            <dependencies>
+                              <dependency>
+                                <groupId>commons-beanutils</groupId>
+                                <artifactId>commons-beanutils</artifactId>
+                                <version>1.7.0</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-collections</groupId>
+                                <artifactId>commons-collections</artifactId>
+                                <version>3.2.1</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-configuration</groupId>
+                                <artifactId>commons-configuration</artifactId>
+                                <version>1.6</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-digester</groupId>
+                                <artifactId>commons-digester</artifactId>
+                                <version>1.8</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-lang</groupId>
+                                <artifactId>commons-lang</artifactId>
+                                <version>2.4</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-logging</groupId>
+                                <artifactId>commons-logging</artifactId>
+                                <version>1.1.1</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>commons-beanutils</groupId>
+                                <artifactId>commons-beanutils-core</artifactId>
+                                <version>1.8.0</version>
+                              </dependency>
+                              <dependency>
+                                <groupId>manual</groupId>
+                                <artifactId>dep</artifactId>
+                                <version>1</version>
+                              </dependency>
+                            </dependencies>
+                          </dependencyManagement>
+                          <description>A demonstration of maven POM customization</description>
+                        </project>
                 '''))
                 .ignoreWhitespace()
                 .build()
 
         then:
         !diff.hasDifferences()
+    }
+
+    @Issue('#49')
+    def 'substituted dependencies do not have recommendations applied'() {
+        when:
+        def project = ProjectBuilder.builder().build();
+        project.apply plugin: 'java'
+        project.apply plugin: DependencyRecommendationsPlugin
+
+        def recommendations = createFile('recommendations.properties')
+        recommendations << 'com.google.collections:google-collections = 1.0'
+
+        project.dependencyRecommendations {
+            propertiesFile name: 'props', file: recommendations
+        }
+
+        project.repositories {
+            mavenCentral()
+        }
+
+        project.configurations.all {
+            resolutionStrategy.dependencySubstitution {
+                substitute module('com.google.collections:google-collections') with module ('com.google.guava:guava:12.0')
+            }
+        }
+
+        project.dependencies {
+            compile 'com.google.collections:google-collections'
+        }
+
+        def resolutionResult = project.configurations.compile.incoming.resolutionResult
+        def guava = resolutionResult.allDependencies.first()
+
+        then:
+        guava instanceof ResolvedDependencyResult
+        def moduleVersion = (guava as ResolvedDependencyResult).selected.moduleVersion
+        moduleVersion.group == 'com.google.guava'
+        moduleVersion.name == 'guava'
+        moduleVersion.version == '12.0'
     }
 }